backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Hoang Le"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
pydantic = "^2.9.2"
langchain-core = "^0.3.17"
langchain-openai = "^0.2.8"
langgraph = "^0.2.49"
python-dotenv = "^1.0.1"
assistant-stream = "^0.0.5"
# NEW: Tool dependencies
httpx = "^0.27.2"
google-api-python-client = "^2.121.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.2.0"
PyJWT = "^2.8.0"
pydantic-settings = "^2.10.1"
# NEW: Supabase for secure token storage
supabase = "^2.12.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
